chore: remove duplication
