+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/fileresource/JCloudsFileResourceContentStore.java
@@ -27,9 +27,6 @@
  */
 package org.hisp.dhis.fileresource;
 
-import static com.google.common.base.Preconditions.checkNotNull;
-import static org.jclouds.Constants.PROPERTY_ENDPOINT;
-
 import com.google.common.hash.HashCode;
 import com.google.common.hash.Hashing;
 import java.io.File;
@@ -38,40 +35,21 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.URI;
 import java.nio.file.Files;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Properties;
-import java.util.regex.Pattern;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
-import javax.annotation.PostConstruct;
-import javax.annotation.PreDestroy;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.tuple.Pair;
-import org.hisp.dhis.external.conf.ConfigurationKey;
-import org.hisp.dhis.external.conf.DhisConfigurationProvider;
-import org.hisp.dhis.external.location.LocationManager;
-import org.jclouds.ContextBuilder;
+import org.hisp.dhis.jclouds.JCloudsStore;
 import org.jclouds.blobstore.BlobRequestSigner;
-import org.jclouds.blobstore.BlobStore;
-import org.jclouds.blobstore.BlobStoreContext;
 import org.jclouds.blobstore.ContainerNotFoundException;
 import org.jclouds.blobstore.LocalBlobRequestSigner;
 import org.jclouds.blobstore.domain.Blob;
 import org.jclouds.blobstore.internal.RequestSigningUnsupported;
-import org.jclouds.domain.Credentials;
-import org.jclouds.domain.Location;
-import org.jclouds.domain.LocationBuilder;
-import org.jclouds.domain.LocationScope;
-import org.jclouds.filesystem.reference.FilesystemConstants;
 import org.jclouds.http.HttpRequest;
-import org.jclouds.http.HttpResponseException;
-import org.jclouds.rest.AuthorizationException;
-import org.jclouds.s3.reference.S3Constants;
 import org.joda.time.Minutes;
 import org.springframework.stereotype.Service;
 
@@ -79,127 +57,13 @@ import org.springframework.stereotype.Service;
  * @author Halvdan Hoem Grelland
  */
 @Slf4j
+@RequiredArgsConstructor
 @Service("org.hisp.dhis.fileresource.FileResourceContentStore")
 public class JCloudsFileResourceContentStore implements FileResourceContentStore {
-  private static final Pattern CONTAINER_NAME_PATTERN =
-      Pattern.compile("^(?![.-])(?=.{1,63}$)([.-]?[a-zA-Z0-9]+)+$");
-
   private static final long FIVE_MINUTES_IN_SECONDS =
       Minutes.minutes(5).toStandardDuration().getStandardSeconds();
 
-  private BlobStore blobStore;
-
-  private BlobStoreContext blobStoreContext;
-
-  private BlobStoreProperties config;
-
-  // -------------------------------------------------------------------------
-  // Providers
-  // -------------------------------------------------------------------------
-
-  private static final String JCLOUDS_PROVIDER_KEY_FILESYSTEM = "filesystem";
-
-  private static final String JCLOUDS_PROVIDER_KEY_AWS_S3 = "aws-s3";
-
-  private static final String JCLOUDS_PROVIDER_KEY_TRANSIENT = "transient";
-
-  private static final List<String> SUPPORTED_PROVIDERS =
-      Arrays.asList(
-          JCLOUDS_PROVIDER_KEY_FILESYSTEM,
-          JCLOUDS_PROVIDER_KEY_AWS_S3,
-          JCLOUDS_PROVIDER_KEY_TRANSIENT);
-
-  // -------------------------------------------------------------------------
-  // Dependencies
-  // -------------------------------------------------------------------------
-
-  private final LocationManager locationManager;
-
-  private final DhisConfigurationProvider configurationProvider;
-
-  public JCloudsFileResourceContentStore(
-      LocationManager locationManager, DhisConfigurationProvider configurationProvider) {
-    checkNotNull(locationManager);
-    checkNotNull(configurationProvider);
-
-    this.locationManager = locationManager;
-    this.configurationProvider = configurationProvider;
-  }
-
-  // -------------------------------------------------------------------------
-  // Life cycle management
-  // -------------------------------------------------------------------------
-
-  @PostConstruct
-  public void init() {
-    // ---------------------------------------------------------------------
-    // Bootstrap config
-    // ---------------------------------------------------------------------
-
-    config =
-        new BlobStoreProperties(
-            configurationProvider.getProperty(ConfigurationKey.FILESTORE_PROVIDER),
-            configurationProvider.getProperty(ConfigurationKey.FILESTORE_LOCATION),
-            configurationProvider.getProperty(ConfigurationKey.FILESTORE_ENDPOINT),
-            configurationProvider.getProperty(ConfigurationKey.FILESTORE_CONTAINER));
-
-    Pair<Credentials, Properties> providerConfig =
-        configureForProvider(
-            config,
-            configurationProvider.getProperty(ConfigurationKey.FILESTORE_IDENTITY),
-            configurationProvider.getProperty(ConfigurationKey.FILESTORE_SECRET));
-
-    // ---------------------------------------------------------------------
-    // Set up JClouds context
-    // ---------------------------------------------------------------------
-
-    blobStoreContext =
-        ContextBuilder.newBuilder(config.provider)
-            .credentials(providerConfig.getLeft().identity, providerConfig.getLeft().credential)
-            .overrides(providerConfig.getRight())
-            .build(BlobStoreContext.class);
-
-    blobStore = blobStoreContext.getBlobStore();
-
-    Location provider =
-        new LocationBuilder()
-            .scope(LocationScope.PROVIDER)
-            .id(config.provider)
-            .description(config.provider)
-            .build();
-
-    try {
-      blobStore.createContainerInLocation(createRegionLocation(config, provider), config.container);
-
-      log.info(
-          String.format(
-              "File store configured with provider: '%s', container: '%s' and location: '%s'.",
-              config.provider, config.container, config.location));
-    } catch (HttpResponseException ex) {
-      log.error(
-          String.format(
-              "Could not configure file store with provider '%s' and container '%s'.\n"
-                  + "File storage will not be available.",
-              config.provider, config.container),
-          ex);
-    } catch (AuthorizationException ex) {
-      log.error(
-          String.format(
-              "Could not authenticate with file store provider '%s' and container '%s'. "
-                  + "File storage will not be available.",
-              config.provider, config.location),
-          ex);
-    }
-  }
-
-  @PreDestroy
-  public void cleanUp() {
-    blobStoreContext.close();
-  }
-
-  // -------------------------------------------------------------------------
-  // FileResourceContentStore implementation
-  // -------------------------------------------------------------------------
+  private final JCloudsStore jCloudsStore;
 
   @Override
   public InputStream getFileResourceContent(String key) {
@@ -259,7 +123,7 @@ public class JCloudsFileResourceContentStore implements FileResourceContentStore
     }
 
     try {
-      blobStore.putBlob(config.container, blob);
+      jCloudsStore.getBlobStore().putBlob(jCloudsStore.getBlobContainer(), blob);
     } catch (Exception e) {
       log.error("File upload failed: ", e);
       return null;
@@ -300,7 +164,7 @@ public class JCloudsFileResourceContentStore implements FileResourceContentStore
 
       if (blob != null) {
         try {
-          blobStore.putBlob(config.container, blob);
+          jCloudsStore.getBlobStore().putBlob(jCloudsStore.getBlobContainer(), blob);
 
           Files.deleteIfExists(file.toPath());
         } catch (ContainerNotFoundException e) {
@@ -329,7 +193,7 @@ public class JCloudsFileResourceContentStore implements FileResourceContentStore
 
   @Override
   public URI getSignedGetContentUri(String key) {
-    BlobRequestSigner signer = blobStoreContext.getSigner();
+    BlobRequestSigner signer = jCloudsStore.getBlobRequestSigner();
 
     if (!requestSigningSupported(signer)) {
       return null;
@@ -338,7 +202,8 @@ public class JCloudsFileResourceContentStore implements FileResourceContentStore
     HttpRequest httpRequest;
 
     try {
-      httpRequest = signer.signGetBlob(config.container, key, FIVE_MINUTES_IN_SECONDS);
+      httpRequest =
+          signer.signGetBlob(jCloudsStore.getBlobContainer(), key, FIVE_MINUTES_IN_SECONDS);
     } catch (UnsupportedOperationException uoe) {
       return null;
     }
@@ -379,19 +244,21 @@ public class JCloudsFileResourceContentStore implements FileResourceContentStore
   }
 
   private Blob getBlob(String key) {
-    return blobStore.getBlob(config.container, key);
+    return jCloudsStore.getBlobStore().getBlob(jCloudsStore.getBlobContainer(), key);
   }
 
   private boolean blobExists(String key) {
-    return key != null && blobStore.blobExists(config.container, key);
+    return key != null
+        && jCloudsStore.getBlobStore().blobExists(jCloudsStore.getBlobContainer(), key);
   }
 
   private void deleteBlob(String key) {
-    blobStore.removeBlob(config.container, key);
+    jCloudsStore.getBlobStore().removeBlob(jCloudsStore.getBlobContainer(), key);
   }
 
   private Blob createBlob(@Nonnull FileResource fileResource, @Nonnull byte[] bytes) {
-    return blobStore
+    return jCloudsStore
+        .getBlobStore()
         .blobBuilder(fileResource.getStorageKey())
         .payload(bytes)
         .contentLength(bytes.length)
@@ -406,7 +273,8 @@ public class JCloudsFileResourceContentStore implements FileResourceContentStore
       String fileDimension,
       @Nonnull File file,
       @Nonnull String contentMd5) {
-    return blobStore
+    return jCloudsStore
+        .getBlobStore()
         .blobBuilder(StringUtils.join(fileResource.getStorageKey(), fileDimension))
         .payload(file)
         .contentLength(file.length())
@@ -419,102 +287,5 @@ public class JCloudsFileResourceContentStore implements FileResourceContentStore
   private boolean requestSigningSupported(BlobRequestSigner signer) {
     return !(signer instanceof RequestSigningUnsupported)
         && !(signer instanceof LocalBlobRequestSigner);
-  }
-
-  private static Location createRegionLocation(BlobStoreProperties config, Location provider) {
-    return config.location != null
-        ? new LocationBuilder()
-            .scope(LocationScope.REGION)
-            .id(config.location)
-            .description(config.location)
-            .parent(provider)
-            .build()
-        : null;
-  }
-
-  private Pair<Credentials, Properties> configureForProvider(
-      BlobStoreProperties properties, String identity, String secret) {
-    Properties overrides = new Properties();
-    Credentials credentials = new Credentials("Unused", "Unused");
-
-    if (properties.provider.equals(JCLOUDS_PROVIDER_KEY_FILESYSTEM)
-        && locationManager.externalDirectorySet()) {
-      overrides.setProperty(
-          FilesystemConstants.PROPERTY_BASEDIR, locationManager.getExternalDirectoryPath());
-    } else if (properties.provider.equals(JCLOUDS_PROVIDER_KEY_AWS_S3)) {
-      overrides.setProperty(S3Constants.PROPERTY_S3_VIRTUAL_HOST_BUCKETS, "false");
-
-      if (!properties.endpoint.isEmpty()) {
-        overrides.setProperty(PROPERTY_ENDPOINT, properties.endpoint);
-      }
-
-      credentials = new Credentials(identity, secret);
-      if (credentials.identity.isEmpty() || credentials.credential.isEmpty()) {
-        log.warn("AWS S3 store configured without credentials, authentication not possible.");
-      }
-    }
-
-    return Pair.of(credentials, overrides);
-  }
-
-  // -------------------------------------------------------------------------
-  // Internal classes
-  // -------------------------------------------------------------------------
-
-  private class BlobStoreProperties {
-    private String provider;
-
-    private final String location;
-
-    private final String endpoint;
-
-    private String container;
-
-    BlobStoreProperties(String provider, String location, String endpoint, String container) {
-      this.provider = provider;
-      this.location = location;
-      this.endpoint = endpoint;
-      this.container = container;
-
-      validate();
-      validateAndSelectProvider();
-    }
-
-    private void validate() {
-      if (!isValidContainerName(container)) {
-        if (container != null) {
-          log.warn(
-              String.format(
-                  "Container name '%s' is illegal. "
-                      + "Standard domain name naming conventions apply (no underscores allowed). "
-                      + "Using default container name ' %s'",
-                  container, ConfigurationKey.FILESTORE_CONTAINER.getDefaultValue()));
-        }
-
-        container = ConfigurationKey.FILESTORE_CONTAINER.getDefaultValue();
-      }
-    }
-
-    private boolean isValidContainerName(String containerName) {
-      return containerName != null && CONTAINER_NAME_PATTERN.matcher(containerName).matches();
-    }
-
-    private void validateAndSelectProvider() {
-      if (!SUPPORTED_PROVIDERS.contains(provider)) {
-        log.warn(
-            "Ignored unsupported file store provider '"
-                + provider
-                + "', using file system provider.");
-        provider = JCLOUDS_PROVIDER_KEY_FILESYSTEM;
-      }
-
-      if (provider.equals(JCLOUDS_PROVIDER_KEY_FILESYSTEM)
-          && !locationManager.externalDirectorySet()) {
-        log.info(
-            "File system file store provider could not be configured; external directory is not set. "
-                + "Falling back to in-memory provider.");
-        provider = JCLOUDS_PROVIDER_KEY_TRANSIENT;
-      }
-    }
   }
 }
